Move conformance tests to new file
